+++ python/fast_mlsirm/__init__.py
@@ -95,6 +95,7 @@ from .exposure import (
     kl_select as kl_select,
     owen_update as owen_update,
     owen_cat as owen_cat,
+    ccat_select as ccat_select,
 )
 from .rsm import fit_rsm as fit_rsm, RsmFit as RsmFit
 from .mixed import fit_mixed_items as fit_mixed_items, MixedFormatFit as MixedFormatFit, MixedItemParameters as MixedItemParameters
@@ -250,6 +251,7 @@ __all__ = [
     "kl_select",
     "owen_update",
     "owen_cat",
+    "ccat_select",
     "AStratifiedResult",
     "omega_total_1f",
     "omega_total_1f_from_data",